add GraphBodyView

+++ MuscleRecord/ViewModel.swift
@@ -14,7 +14,6 @@ class ViewModel: ObservableObject {
     @Published var records3 = [Record]()
     @Published var records9 = [Record]()
     @Published var records27 = [Record]()
-    @Published var weightArray = [Float]()
     @Published var maxWeight: Float = 0.0
     @Published var latestID: String = ""
     @Published var latestID3: String = ""
@@ -44,7 +43,7 @@ class ViewModel: ObservableObject {
             return Color("ThemeColor0")
         }
     }
-
+    
     func dateFormat(date: Date) -> String {
         let f = DateFormatter()
         f.dateStyle = .long
@@ -54,29 +53,45 @@ class ViewModel: ObservableObject {
     
     func updateEvent(event: Event, newName: String) {
         let db = Firestore.firestore()
-        db.collection("user").document(event.id).updateData(["name": newName])
+        db.collection("user").document(event.id).updateData(["name": newName]) { error in
+            if let error = error {
+                print("updateEvent中のエラー: \(error)")
+            } else {
+                print("updateEvent成功")
+            }
+        }
     }
     
     func deleteEvent(event: Event) {
         let db = Firestore.firestore()
-        db.collection("user").document(event.id).delete() { err in
-            if let err = err {
-                print("Error removing document: \(err)")
+        db.collection("user").document(event.id).delete() { error in
+            if let error = error {
+                print("deleteEvent中のエラー: \(error)")
             } else {
-                print("Document successfully removed!")
+                print("deleteEvent成功")
             }
         }
     }
     
     func addEvent(_ name: String) {
         let db = Firestore.firestore()
-        db.collection("user").document(name).setData(["latestWeight": 0, "latestRep": 0, "latestDate": Date(timeInterval: -60*60*24, since: .now)])
+        db.collection("user").document(name).setData(["latestWeight": 0, "latestRep": 0, "latestDate": Date(timeInterval: -60*60*24, since: .now)]) { error in
+            if let error = error {
+                print("addEvent中のエラー: \(error)")
+            } else {
+                print("addEvent成功")
+            }
+        }
     }
     
     func getEvent() {
         let db = Firestore.firestore()
-            db.collection("user").getDocuments { snapshot, error in
-                if let snapshot = snapshot{
+        db.collection("user").getDocuments { snapshot, error in
+            if let snapshot = snapshot{
+                if let error = error {
+                    print("getEvent中のエラー: \(error)")
+                } else {
+                    print("getEvent成功")
                     DispatchQueue.main.async {
                         self.events = snapshot.documents.map { d in
                             let timeStamp = d["latestDate"] as! Timestamp
@@ -85,119 +100,80 @@ class ViewModel: ObservableObject {
                     }
                 }
             }
+        }
     }
     
     func getRecord(event: Event) {
         let db = Firestore.firestore()
         db.collection("user").document(event.id).collection("records").order(by: "date").getDocuments { (snapshot, error) in
             if let snapshot = snapshot{
-                DispatchQueue.main.async {
-                    snapshot.documents.forEach { d in
-                        self.latestID = d.documentID
-                        let timeStamp: Timestamp = d["date"] as? Timestamp ?? Timestamp()
-                        let date = timeStamp.dateValue()
-                        if let newRecord = self.oldRecord {
-                            var dateDifference = (Calendar.current.dateComponents([.day], from: newRecord.date, to: timeStamp.dateValue())).day! - 1
-                            while dateDifference > 0 {
-                                self.records.append(Record(id: UUID().uuidString, date: Date(timeInterval: TimeInterval(-60*60*24*dateDifference), since: date), weight: newRecord.weight, rep: newRecord.rep, dummy: true))
-                                dateDifference -= 1
+                if let error = error {
+                    print("getRecord中のエラー: \(error)")
+                } else {
+                    print("getRecord成功")
+                    DispatchQueue.main.async {
+                        snapshot.documents.forEach { d in
+                            self.latestID = d.documentID
+                            let timeStamp: Timestamp = d["date"] as? Timestamp ?? Timestamp()
+                            let date = timeStamp.dateValue()
+                            if let newRecord = self.oldRecord {
+                                var dateDifference = (Calendar.current.dateComponents([.day], from: newRecord.date, to: timeStamp.dateValue())).day! - 1
+                                while dateDifference > 0 {
+                                    self.records.append(Record(id: UUID().uuidString, date: Date(timeInterval: TimeInterval(-60*60*24*dateDifference), since: date), weight: newRecord.weight, rep: 0, dummy: true))
+                                    dateDifference -= 1
+                                }
+                            }
+                            let record = Record(id: d.documentID, date: date, weight: d["weight"] as? Float ?? 0, rep: d["rep"] as? Int ?? 0, dummy: false)
+                            self.oldRecord = record
+                            self.records.append(record)
+                        }
+                        
+                        for i in 1..<4{
+                            var dataCount = 0
+                            var totalWeight: Float = 0
+                            var totalRep = 0
+                            let period = Int(truncating: pow(3, i) as NSNumber)
+                            let remainder = self.records.count % period
+                            var created = false
+                            self.records.forEach { record in
+                                totalWeight += record.weight
+                                totalRep += record.rep
+                                dataCount += 1
+                                if remainder != 0 && remainder == dataCount && created == false {
+                                    let recordID = UUID().uuidString
+                                    self.records3.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(dataCount), rep: totalRep/dataCount, dummy: false))
+                                    self.latestID3 = recordID
+                                    created = true
+                                    totalWeight = 0
+                                    totalRep = 0
+                                    dataCount = 0
+                                }
+                                if dataCount == period {
+                                    let recordID = UUID().uuidString
+                                    if period == 3 {
+                                        self.records3.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
+                                        self.latestID3 = recordID
+                                    } else if period == 9 {
+                                        self.records9.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
+                                        self.latestID9 = recordID
+                                    } else if period == 27 {
+                                        self.records27.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(period), rep: totalRep/period, dummy: false))
+                                        self.latestID27 = recordID
+                                    }
+                                    totalWeight = 0
+                                    totalRep = 0
+                                    dataCount = 0
+                                }
                             }
                         }
-                        let record = Record(id: d.documentID, date: date, weight: d["weight"] as? Float ?? 0, rep: d["rep"] as? Int ?? 0, dummy: false)
-                        self.oldRecord = record
-                        self.records.append(record)
-                    }
-                    
-                    var dataCount = 0
-                    var totalWeight: Float = 0
-                    var totalRep = 0
-                    var remainder = self.records.count % 3
-                    var created = false
-                    self.records.forEach { record in
-                        totalWeight += record.weight
-                        totalRep += record.rep
-                        dataCount += 1
-                        if remainder != 0 && remainder == dataCount && created == false {
-                            let recordID = UUID().uuidString
-                            self.records3.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(dataCount), rep: totalRep/dataCount, dummy: false))
-                            self.latestID3 = recordID
-                            created = true
-                            totalWeight = 0
-                            totalRep = 0
-                            dataCount = 0
-                        }
-                        if dataCount == 3 {
-                            let recordID = UUID().uuidString
-                            self.records3.append(Record(id: recordID, date: record.date, weight: totalWeight/3, rep: totalRep/3, dummy: false))
-                            self.latestID3 = recordID
-                            totalWeight = 0
-                            totalRep = 0
-                            dataCount = 0
+                        
+                        snapshot.documents.forEach { d in
+                            if self.maxWeight < d["weight"] as? Float ?? 0.0 {
+                                self.maxWeight = d["weight"] as? Float ?? 0.0
+                            }
                         }
                     }
-                    
-                    dataCount = 0
-                    totalWeight = 0
-                    totalRep = 0
-                    remainder = self.records.count % 9
-                    created = false
-                    self.records.forEach { record in
-                        totalWeight += record.weight
-                        totalRep += record.rep
-                        dataCount += 1
-                        if remainder != 0 && remainder == dataCount && created == false {
-                            let recordID = UUID().uuidString
-                            self.records9.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(dataCount), rep: totalRep/dataCount, dummy: false))
-                            self.latestID9 = recordID
-                            created = true
-                            totalWeight = 0
-                            totalRep = 0
-                            dataCount = 0
-                        }
-                        if dataCount == 9 {
-                            let recordID = UUID().uuidString
-                            self.records9.append(Record(id: recordID, date: record.date, weight: totalWeight/9, rep: totalRep/9, dummy: false))
-                            self.latestID9 = recordID
-                            totalWeight = 0
-                            totalRep = 0
-                            dataCount = 0
-                        }
-                    }
-                    
-                    dataCount = 0
-                    totalWeight = 0
-                    totalRep = 0
-                    remainder = self.records.count % 27
-                    created = false
-                    self.records.forEach { record in
-                        totalWeight += record.weight
-                        totalRep += record.rep
-                        dataCount += 1
-                        if remainder != 0 && remainder == dataCount && created == false {
-                            let recordID = UUID().uuidString
-                            self.records27.append(Record(id: recordID, date: record.date, weight: totalWeight/Float(dataCount), rep: totalRep/dataCount, dummy: false))
-                            self.latestID27 = recordID
-                            created = true
-                            totalWeight = 0
-                            totalRep = 0
-                            dataCount = 0
-                        }
-                        if dataCount == 27 {
-                            let recordID = UUID().uuidString
-                            self.records27.append(Record(id: recordID, date: record.date, weight: totalWeight/27, rep: totalRep/27, dummy: false))
-                            self.latestID27 = recordID
-                            totalWeight = 0
-                            totalRep = 0
-                            dataCount = 0
-                        }
-                    }
-                    
-                    self.weightArray = snapshot.documents.map { d in
-                        if self.maxWeight < d["weight"] as? Float ?? 0.0 {
-                            self.maxWeight = d["weight"] as? Float ?? 0.0
-                        }
-                        return d["weight"] as? Float ?? 0.0
-                    }
+
                 }
             }
         }
@@ -205,20 +181,49 @@ class ViewModel: ObservableObject {
     
     func addRecord(event: Event, weight: Float, rep: Int) {
         let db = Firestore.firestore()
-        db.collection("user").document(event.id).setData(["latestWeight": weight, "latestRep": rep, "latestDate": Date()])
-        db.collection("user").document(event.id).collection("records").addDocument(data: ["date": Date(), "weight": weight, "rep": rep])
+        db.collection("user").document(event.id).setData(["latestWeight": weight, "latestRep": rep, "latestDate": Date()]) { error in
+            if let error = error {
+                print("addRecord(event)中のエラー: \(error)")
+            } else {
+                print("addRecord(event)成功")
+            }
+        }
+        db.collection("user").document(event.id).collection("records").addDocument(data: ["date": Date(), "weight": weight, "rep": rep]) { error in
+            if let error = error {
+                print("addRecord(record)中のエラー: \(error)")
+            } else {
+                print("addRecord(record)成功")
+            }
+        }
     }
     
     func updateRecord(event: Event, weight: Float, rep: Int) {
         let db = Firestore.firestore()
         db.collection("user").document(event.id).collection("records").order(by: "date", descending: true).limit(to: 1).getDocuments { snapshot, error in
             if let snapshot = snapshot {
-                snapshot.documents.forEach { d in
-                    let latestRecordID = d.documentID
-                    db.collection("user").document(event.id).collection("records").document(latestRecordID).delete()
+                if let error = error {
+                    print("updateRecord(delete)中のエラー: \(error)")
+                } else {
+                    print("updateRecord(delete)成功")
+                    snapshot.documents.forEach { d in
+                        let latestRecordID = d.documentID
+                        db.collection("user").document(event.id).collection("records").document(latestRecordID).delete()
+                    }
                 }
-                db.collection("user").document(event.id).setData(["latestWeight": weight, "latestRep": rep,"latestDate": Date()])
-                db.collection("user").document(event.id).collection("records").addDocument(data: ["date": Date(), "weight": weight, "rep": rep])
+            }
+            db.collection("user").document(event.id).setData(["latestWeight": weight, "latestRep": rep,"latestDate": Date()]) { error in
+                if let error = error {
+                    print("updateRecord(event)中のエラー: \(error)")
+                } else {
+                    print("updateRecord(event)成功")
+                }
+            }
+            db.collection("user").document(event.id).collection("records").addDocument(data: ["date": Date(), "weight": weight, "rep": rep]) { error in
+                if let error = error {
+                    print("updateRecord(record)中のエラー: \(error)")
+                } else {
+                    print("updateRecord(record)成功")
+                }
             }
         }
     }
